src/vertex-search/requirements.txt
functions-framework==3.*
firebase-admin==6.2.*
firebase-functions==0.1.*
google-cloud-discoveryengine==0.11.1